% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpi_mlr.R
\name{cpi_mlr}
\alias{cpi_mlr}
\title{Conditional Predictive Impact (CPI) for mlr.}
\usage{
cpi_mlr(task, learner, resampling = NULL, test_data = NULL,
  measure = NULL, test = "t", permute = TRUE, log = TRUE,
  B = 10000, alpha = 0.05, verbose = FALSE, cores = 1)
}
\arguments{
\item{task}{The prediction task.}

\item{learner}{The learner. If you pass a string the learner will be created via \link{makeLearner}.}

\item{resampling}{Resampling description object, name of resampling strategy, "oob" (out-of-bag) or "none" (in-sample loss).}

\item{test_data}{External validation data, use instead of resampling.}

\item{measure}{Performance measure.}

\item{test}{Statistical test to perform, either "t" (t-test) or "fisher" (Fisher permuation test).}

\item{permute}{Permute the feature of interest. Set to \code{FALSE} to drop the feature of interest.}

\item{log}{Set to \code{TRUE} for multiplicative CPI (\eqn{\lambda}), to \code{FALSE} for additive CPI (\eqn{\Delta}).}

\item{B}{Number of permutations for Fisher permutation test.}

\item{alpha}{Significance level for confidence intervals.}

\item{verbose}{Verbose output of resampling procedure.}

\item{cores}{Number CPU cores used.}
}
\value{
\code{data.frame} with a row for each feature and columns:
  \item{Variable}{Variable name}
  \item{CPI}{CPI value}
  \item{SE}{Standard error}
  \item{statistic}{Test statistic (t-test only)}
  \item{p.value}{p-value}
  \item{ci.lo}{Lower limit of confidence interval}
}
\description{
Conditional Predictive Impact (CPI) for mlr.
}
\examples{
library(mlr)
# Regression with linear model
cpi_mlr(task = bh.task, learner = makeLearner("regr.lm"), 
        resampling = makeResampleDesc("Holdout"))

# Classification with logistic regression, log-loss and subsampling
cpi_mlr(task = iris.task, 
        learner = makeLearner("classif.glmnet", predict.type = "prob"), 
        resampling = makeResampleDesc("CV", iters = 5), 
        measure = "logloss", test = "t")
 
# Random forest with out-of-bag error               
cpi_mlr(task = bh.task, learner = makeLearner("regr.ranger", num.trees = 50), 
        resampling = "oob", measure = "mse", test = "t")
        
}
